usb protection: remove comment about a USBGuard bug
The bug pertains to searching for rules with a label.
When the USB protection component starts, it checks whether
a "allow all" rule is present and in case no such rule is found,
it adds "allow all" with a custom label.
When searching for the rule, we could add our label but it does not
matter for the functionality where the rule comes from.
Especially because we neither change nor delete the rule.