plugins: fix linkage of GResources
The libcommon_dep gresources were not ending up in the final gsd-* plugin
binaries. What this means is that the work to disable the GTK theme parsing
at startup was innefective, wasting significant time and memory during
application startup.

Collectively, these are still slowing down login.

Calling g_resources_register() at startup is enough to ensure that the
the empty gtk.css ends up in the target binary.

You can verify the resources show up in the resulting binary with:

  $ gresources list /path/to/gsd-media-keys
  /org/gtk/libgtk/theme/Disabled/gtk.css

This was found with Sysprof while profiling login on GNOME OS.