sharing: Give assigned services 10 seconds before killing them
It's useful to wait a bit before killing an assigned service when
the associated system service goes away.

This avoids having to tear down the user service if the system
service is just getting restarted.

This commit adds a 10 second countdown before slaying the user
service when it's system service counterpart goes down.