sharing: Start grd --handover if grd --system is running
In order to provide a login screen over gnome-remote-desktop,
their needs to be both system and session instance of
gnome-remote-desktop. The latter should get run if the former
is running, and not otherwise.

This commit adds a mechanism for doing this by introducing the
concept of "assigned services" (versus the already existing
configurable services).

Assigned services don't start based on GSettings configuration,
but instead just start if an associated system service is running.