Use the proper hint for setting images for notifications
The icon set via `notify_notification_new()` is used as app-icon in the
DBus request. Unfortunately libnotify doesn't document this behavior. It
didn't really matter since in GNOME Shell we used to display only one
icon. After introducing a header [1] for notification in GNOME Shell we
display
the app-icon and when set the image from the image hint[2].

[1]: https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3103
[2]: https://specifications.freedesktop.org/notification-spec/notification-spec-latest.html#icons-and-images-formats