smartcard: Fetch GckModule list on our own
gck returns the p11-kit proxy modules in its pkcs11 module list.
This is right for many applications, but causes a problem for
g-s-d because it the proxy modules doesn't support block operation
which significantly affects battery life.

gsd is never opening a session with the smartcard or anything like
that, just listening for events, so it's probably okay to use the
underlying modules directly instead of the proxy module.

This commit changes gsd-smartcard to construct the list of GckModules
itself, instead, based on the underlying modules, rather than have
gck give us a list of proxy modules.