tests: Run tests in headless Wayland session
This removes the dependency on Xvfb, and decreases the risk of zombie
X11 frame processes due to SIGKILL being used for terminating mutter.

For xsettings tests, which needs an X11 server, we rely on Xwayland
started by mutter. In order for gsd-xsettings to be able to connect, the
X11 display name and Xauthority path needs to be extracted, which is
done with the help of a named pipe and a simple shell script launched by
mutter.