keyboard: Do not load xkb-options from localed when explicitly set to empty
If a user on a system with non-empty `xkb-options` in `localed` decides
to disable the options locally by setting the GSettings key to an empty
array, the value will be reset back to `localed` value on next log-in.

Let’s stop overriding `xkb-options` if the GSettings key is already set.

A consequence of this patch is that when an administrator changes the
previously empty `xkb-options` in `localed`, users that had already
logged in before the change will keep the empty option list.
But one would likewise not expect `create_initial_settings` to update
the GSettings key on change from a non-empty `xkb-options`
so this patch just makes the behaviour more consistent.