power/backlight: Make backlight backend explicit
Mutter changes the backlight property on VT switches and monitor
reconfigurations, to tell us both whether there are any controllable
backlight, or if there is any backlight to potentially control from
somewhere else. If we control the backlight via e.g. sysfs/udev,
such a signal emission would inadvertently reset the backlight because
we'd think we went from having controllable backlight to not having it,
which isn't true if that backlight was controlled via sysfs.

Handle this by making the backend in control of the backlight explicit,
and only attempt to reset the backlight fields on mutter signal
notifications, if the mutter backlight backend is used.

Closes: https://gitlab.gnome.org/GNOME/gnome-settings-daemon/-/issues/818