media-keys: Allow running custom command in lockscreen
Allow running a custom command in the lockscreen (and everywhere else).
This is useful to allow users to set "background" scripts to run even
when the lockscreen is up, such as being able to control a smart light,
smart plug or any other automation even when the computer is locked.