build: Allow simple usage as subproject
Only useful for gnome-control-center really, but it is handy.

Does require handling of the `gsd-enums.h` directory, because
otherwise it can't be matched to the install directory easily.
This is a Meson limitation, see
https://github.com/mesonbuild/meson/issues/2320.

We can now remove data_inc as well, and have interal users of
gsd-enums.h use the same dependency too, with the same include path as
external users.