smartcard/manager: Handle the case in which we've an invalid token
Apparently a module may tell us that a token is present but then such
token is invalid, so instead of crashing in such case, let's instead
ensure that if we've a token, then it has also to be a valid one.

Without relying on the assumption that the flag is enough.

Closes: #795