smartcard/manager: Consider modules with no slots as removable devices
Some smartcard modules support removable slots but they don't advertise
them until a smartcard device is actually plugged.

In such case we can't know preventively if the module provides
removable-token (e.g. smartcard) support, so we can just assume that it
does or we won't be able to catch the token insertion when this happens.

Closes: #782