data: Add schema for global shortcuts
Adopt this piece of data so far, even though gnome-settings-daemon is
not directly involved with it. At least until we move handling of
shortcuts altogether to gsettings-desktop-schemas.

Since the format for shortcut IDs is undefined and might not be able
to be stored as DConf path components, avoid the temptation to unroll
the configuration, and store settings as a "a(sa{sv})" containing
the configuration for all shortcuts. This GVariant format coincides
with the response format that xdp-gnome expects from the UI.