plugins: Drop Wacom plugin
Many years back, this plugin used to do a lot of the heavy lifting
in configuring Wacom devices on a X11 environment. With Wayland
coming much of this configuration handling was lifted to Mutter
instead, leaving the Wacom plugin only in charge of two things:

- Migrating settings from older locations/formats to newer ones,
  something not exercised since 2018.
- Handling the small OLED displays alongside pad buttons in the
  Wacom Intuos 4 set of models. This series is already discontinued
  and the gimmick was not attempted again in any other models.

We cannot say that the first role will not ever be needed again,
but the GsdSettingsMigrate* infrastructure remains and can be
called from other less dedicated places.

The second role implies a pkexec executable that for the most
part, cannot be tested. This is a liability nowadays for marginal
gain. The tablet will still be supported, the OLEDs will just
remain blank.