power: Mock mutter with a custom dbusmock template
The mutter we spawn for the tests is only used for Xsettings which is
hard to mock. The rest of the APIs we need from mutter are available as
dbus interfaces that we can mock.

For now there isn't a lot of benefit, but in the next commit we will
start to rely on the mutter backlight API and having it mocked means we
don't have to figure out how to get backlights on a headless mutter
instance.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-settings-daemon/-/merge_requests/418>