common: Port to gtk4
The few plugins using gtk mainly use GDK api, with the exception
of print notifications in order to find out the default paper size.
And all of them do use API that did not change in the 3.x->4.x
transition. The same do the libgnome-desktop users.

As a side effect of the transition, we can no longer naively modify
the GTK theme, so stop trying to do so.