media-keys: Use GAppLaunchContext directly
The only advantage we might ever obtain from GDK here is integrating
with startup notification for applications being launcher. However,
we do just receive a timestamp, which is by itself not sufficient to
integrate with startup notification on Wayland.

While we think that through, using GTK has other side effects, so
it's better to avoid GdkAppLaunchContext here altogether.