a11y-settings: Manage Orca's lifecycle
Previously, Orca was launched by gnome-session via a .desktop file. The
desktop file instructed gnome-session to monitor the
screen-reader-enabled gsetting and run Orca for as long as it's true.

Since GNOME 49, gnome-session doesn't use .desktop files to launch
session services anymore, and has removed support for autostart
conditions. Thus, gnome-session no longer watches the
screen-reader-enabled gsetting.

Luckily, gsd-a11y-settings monitors the gsetting, so this is a natural
place for this logic. Note that we plan to move Orca into a systemd
service, however it's not quite ready for this transition [1]

[1]: https://gitlab.gnome.org/GNOME/orca/-/issues/619