a11y-settings: Manage Orca's lifecycle
Previously, Orca was launched by gnome-session via a .desktop file. The
desktop file instructed gnome-session to monitor the
screen-reader-enabled gsetting and run Orca for as long as it's true.

Since GNOME 49, gnome-session doesn't use .desktop files to launch
session services anymore, and has removed support for autostart
conditions. Thus, gnome-session no longer watches the
screen-reader-enabled gsetting.

Luckily, gsd-a11y-settings monitors the gsetting, so this is a natural
place for this logic.

Note that we call this codepath the "fallback" because a future commit
will introduce a codepath that uses systemd to launch Orca, and the code
introduced here will be used only when systemd fails