power: Fix boolean property value extraction
The code here is expecting a "(b)" GVariant signature, which does
not match with "b" as sent by GNOME Shell. This produces a barrage of
warnings like:

  gsd-power[11982]: the GVariant format string '(b)' has a type of '(b)' but the given value has a type of 'b'
  gsd-power[11982]: g_variant_get: assertion 'valid_format_string (format_string, TRUE, value)' failed

And other side effects. Fix this boolean value extraction, to fix
the warnings and detect correctly that GNOME Shell is in charge of
backlight brightness.

Fixes: f7555c02 ("power: Add new dimming and auto brightness API based on Shell.Brightness")