power: Rework ambient light sensor normalization
This reworks the normalization and fixes two things in particular:

The percentage `pc` is an int so if we want values different
from `0` and `1` we need to use floating point division.

When gsd-power controlled the backlight it would init that value from
the backlight on startup but we don't have that available anymore so
can only pick an arbitrary value.

Co-authored-by: Guido Günther <agx@sigxcpu.org>
Fixes: 08ae6461 ("power: Switch from direct mutter backlight handling to Shell.Brightness")