housekeeping: Don't mark donation reminder notification as transient
When the notification was still marked as critical prior to 54b5d05e,
setting the transient hint would ensure that it was not showing up in
the calendar menu, which made sense because it was shown on screen until
manually closed or interacted with. However now that it is not critical
anymore, this means the notification disappears after being hovered and
unhovered, without any other user interaction.

Fixes: 54b5d05e ("housekeeping: Downgrade donation reminder urgency to "normal"")