power: Trigger re-normalization on out-of-band brightness changes
When brightness is auto-managed by ALS, we still have two situations
where the brightness changes do not come from it:

- User changes to the display brightness, either through keybindings,
  or the sliders in the shell menu being changed.
- Dimming requests on user inactivity.

The response to both should be similar, the abrupt brightness change
can alter ALS readings, so the brightness resulting of the change must
be used as a new baseline for normalization.

Tracking of user-triggered changes is done through a new BrightnessChanged
signal in the org.gnome.Shell.Brightness D-Bus interface. This will do
nothing on older versions of GNOME Shell without the signal, so effectively
brightness changes would not trigger a re-normalization there, as before.
If the signal exists and is emitted, then it will be used.